// auctionhouseapp.client/src/dto/liveAuction/ILiveAuctionPreviewResponse.d.ts
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

import { ILiveAuctionItem } from './ILiveAuctionItem';

export interface ILiveAuctionPreviewResponse
{
	items: ILiveAuctionItem[];
}