// auctionhouseapp.client/src/dto/liveAuction/ILiveAuctionStatusResponse.d.ts
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

import { ILiveAuctionStatus } from './ILiveAuctionStatus';

export interface ILiveAuctionStatusResponse
{
	data: ILiveAuctionStatus;
}