// auctionhouseapp.client/src/dto/liveAuction/IRecordBidResponse.d.ts
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

export interface IRecordBidResponse
{
	bidId: number;
	itemId: string;
	paddleNum: string;
	paddleName: string;
	bidAmount: number;
	timestamp: string;
}