// auctionhouseapp.client/src/dto/liveAuction/IHammerResponse.d.ts
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

export interface IHammerResponse
{
	itemId: string;
	auctionResult: string;
	winnerPaddleNum: string;
	winnerName: string;
	hammerPrice: number;
}